moved faculty and mentor into components
